Added CustomUserIO Module
## What?
Rewrote Exit-OnKeyPress to be an independent module separate from the main script
## Why?
Exiting the shell interface given user input is a prime function to be reused in other scripts. Creating this module allows portability of the Exit-OnKeyPress code and version control.
## How?
Rewrote the Exit-OnKeyPress into its own module. The function was updated with in-file documentation.

Additionally, some new functionality was implemented to increase portability. Namely, the function now allows deletion of the message and newline characters which were previously always included. New usability was added by allowing function calls to format the text after the default output.
## Testing?
Verified to import and work without error when dot sourced into the main script.

## Anything Else?
A module manifest should be created to ensure the module version stays locked.